aria-interviewer/backend/requirements.txt
fastapi
uvicorn
groq
pdfplumber
python-multipart
python-dotenv
pydantic
httpx
supabase
